backend/requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-jwt-extended==4.5.3
flask-cors==4.0.0
flask-mail==0.9.1
flask-login==0.6.2
python-dotenv==1.0.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
pytest==7.4.2
black==23.7.0
flake8==6.1.0
python-jose==3.3.0
requests==2.31.0
pillow==10.0.0 